% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.data.frame.replacement_list}
\alias{as.data.frame.replacement_list}
\title{Convert a replacement_list object into a data.frame.}
\usage{
\method{as.data.frame}{replacement_list}(x)
}
\arguments{
\item{x}{A replacement_list.}
}
\value{
A data.frame. Each row corresponds to an entry in the
  replacement_list.
}
\description{
Convert a replacement_list object into a data.frame.
}

