% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.matrix.replacement_list}
\alias{as.matrix.replacement_list}
\title{Convert a replacement_list object into a character matrix.}
\usage{
\method{as.matrix}{replacement_list}(x)
}
\arguments{
\item{x}{A replacement_list.}
}
\value{
A character matrix. Each row corresponds to an entry in the
  replacement_list.
}
\description{
Convert a replacement_list object into a character matrix.
}

