% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPatternMatches.R
\name{countPatternMatches}
\alias{countPatternMatches}
\title{Count how many matches are generated by each pattern.}
\usage{
countPatternMatches(X, patterns)
}
\arguments{
\item{X}{A character vector where matches are sought.}

\item{patterns}{A vector of regular expression patterns.}
}
\value{
A integer vector of same length as \code{patterns} with a count of 
  matches.
}
\description{
Count how many matches are generated by each pattern.
}
\details{
This is meant to be useful when developing a vector of regexes to 
  apply to text. Unlike \code{patternappy()}, which only finds the first 
  regex that matches each element of text, a text element here can match 
  multiple regexes. Therefore, the sum of the returned vector may not equal 
  the length of the input \code{X}.
}

