% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPatternMatches.R
\name{filterPatternMatches}
\alias{filterPatternMatches}
\title{Filter out all of the records that match an existing pattern.}
\usage{
filterPatternMatches(X, patterns)
}
\arguments{
\item{X}{A character vector where matches are sought.}

\item{patterns}{A vector of regular expression patterns.}
}
\value{
The subset of \code{X} that did not match any of the regexes in 
  \code{patterns}
}
\description{
Filter out all of the records that match an existing pattern.
}
\details{
This is meant to be useful when developing a vector of regexes to 
  apply to text. Regular expressions can be appended to \code{patterns}
  interactively until the edge cases are all covered, and then
  \code{patternapply()} can be deployed to extract data.
}

