% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternapply.R
\docType{package}
\name{patternapply-package}
\alias{patternapply-package}
\title{patternappy: A package for attempting to apply patterns to a vector.}
\description{
This package contains the function \code{patternapply()}, which iteratively 
tries to match a list of regular expressions to a vector and returns the
associated replacement vector. This is useful for turning a column of text
data with different formats into a matrix or data frame. S3 generics are also
provided to assist with this.
}

