% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternapply.R
\name{patternapply}
\alias{patternapply}
\title{Iteratively try patterns against a character vector.}
\usage{
patternapply(X, patterns, replacements = list(paste(seq_along(patterns))))
}
\arguments{
\item{X}{A character vector where matches are sought.}

\item{patterns}{A vector of regular expression patterns.}

\item{replacements}{A list of replacement information, which must match the 
length of \code{patterns}. Each element must be a character vector. This can 
include backreferences "\1" to "\9" to parenthesized subexpressions of the 
corresponding pattern.}
}
\value{
A list of replacement vectors with class "replacement_list".
}
\description{
Iteratively try patterns against a character vector.
}

